+++ src/lib/listingEvents/index.js
@@ -16,3 +16,9 @@ export {
 } from "./buildListingEventPayload";
 
 export { writeListingEvent, emitListingEventAfterMutation } from "./writeListingEvent";
+
+export {
+  LISTING_EVENT_PRESENTATION,
+  formatListingEventRelativeTime,
+  presentListingEvent,
+} from "./listingEventPresentation";